services/news_monitor/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
psycopg2-binary==2.9.9
requests==2.32.3
feedparser==6.0.11
nltk==3.9.1
textblob==0.18.0
python-multipart==0.0.12
